requirements.txt
cshogi
numpy<2
tqdm
onnxruntime-gpu

# for Google Drive API
google-auth-oauthlib
google-api-python-client
requests
